Login screens: extract some classes
